requirements.txt
langchain==0.3.27
langchain_community==0.3.30
langchain_openai==0.3.33
python-dotenv==1.1.1
